Todos/Findings: alternate way to introduce surface (leverage the bmp stuff)
#####
todos long term:
Pressure boundaries are a bit of a problem in LBM, why do we not use something like the pressure motor from Tareks work to enforce a pressure gradient, would that work.
As for the dangeling of uneven exits or more than one why not use some simple interpolation techniques for that?
Handleling of the nl is suboptimal, we pretty much build them and then throw them away and never use them
#####
Actual:
pretty much still working on the fix for ibm inner and outer, wrote the doxygen for the ibm simulation
I feel like the simulation classes are a bit of spagetti code at some time one will fully break.