Todos/Findings: alternate way to introduce surface (leverage the bmp stuff)
#####
todos long term:
Pressure boundaries are a bit of a problem in LBM, why do we not use something like the pressure motor from Tareks work to enforce a pressure gradient, would that work.
As for the dangeling of uneven exits or more than one why not use some simple interpolation techniques for that?
Handleling of the nl is suboptimal, we pretty much build them and then throw them away and never use them
#####
Actual:
tested and implemented  a bb like link on the last imb outer node so it doesnt dangle
right now they are set to itself so there is no streaming
need to write a test that checks the outer nodes after a while, what happens there ?!