Todos/Findings: alternate way to introduce surface (leverage the bmp stuff)
#####
todos long term:
Pressure boundaries are a bit of a problem in LBM, why do we not use something like the pressure motor from Tareks work to enforce a pressure gradient, would that work.
As for the dangeling of uneven exits or more than one why not use some simple interpolation techniques for that?
Handleling of the nl is suboptimal, we pretty much build them and then throw them away and never use them
#####
Actual:
current problem is that it does a few steps and then stops when the direction changes
debug and do 13 steps that one seems to be the critical one